% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymise.R
\name{anonymise}
\alias{anonymise}
\title{Function to scan a column of free text responses and remove any
obvious references to names}
\usage{
anonymise(data, column, add_names = c(""), extract = TRUE,
  complete = FALSE, identify = FALSE, gender = FALSE)
}
\arguments{
\item{data}{dataframe or tibble with a row per survey response}

\item{column}{name of a character column in the data frame to be anonymised}

\item{add_names}{character vector of names to anonymise (optional)}

\item{extract}{logical indicating whether to look for names to anonymise in the column. Defaults to TRUE.}

\item{complete}{logical indicating whether to anonymise every capitalised word that doesn't start a sentence. Defaults to FALSE.}

\item{identify}{logical indicating whether anonymised names should be distinguishable from each other by numbers. Defaults to FALSE.}

\item{gender}{logical indicating whether to remove references to gender pronouns (he/she him/her etc.). Defaults to FALSE.}
}
\value{
Original dataframe with column anonymised
}
\description{
Function to scan a column of free text responses and remove any
obvious references to names
}
