% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymise.R
\name{anonymise_all}
\alias{anonymise_all}
\title{anonymise_all}
\usage{
anonymise_all(data, add_names = c(""), auto = TRUE, complete = FALSE,
  identify = FALSE, gender = FALSE, quietly = FALSE)
}
\arguments{
\item{data}{dataframe or tibble with a row per survey response}

\item{add_names}{character vector of names to anonymise (optional)}

\item{auto}{logical indicating whether to look for names to anonymise in the column. Defaults to TRUE.}

\item{complete}{logical indicating whether to anonymise every capitalised word that doesn't start a sentence. Defaults to FALSE.}

\item{identify}{logical indicating whether anonymised names should be distinguishable from each other by numbers. Defaults to FALSE.}

\item{gender}{logical indicating whether to remove references to gender pronouns (he/she him/her etc.). Defaults to FALSE.}

\item{quietly}{logical indicating whether to suppress warnings for columns that don't contain names}
}
\value{
Original dataframe with all columns anonymised
}
\description{
Simplify application of anonymise function to every column in the dataset
}
\seealso{
Other anonymiser functions: \code{\link{anonymise_at}},
  \code{\link{anonymise}}
}
\concept{anonymiser functions}
