% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_topline_multiselect.R
\name{apply_topline_multiselect}
\alias{apply_topline_multiselect}
\title{Apply topline generation to a list of data frames}
\usage{
apply_topline_multiselect(list_df, weight, caption, parent)
}
\arguments{
\item{list_df}{A list of data frame objects.}

\item{weight}{A single string of the weighting variable.}

\item{caption}{A list of character vectors of captions for the toplines.}

\item{parent}{A logical vector of booleans indicating whether \code{df} has a parent response column.}
}
\value{
A list containing two elements--- \code{result} and \code{error}, which are lists in and of themselves.
These two lists have the same structure and number of elements. The \code{result} is a list
of list objects. One can think of each element in \code{result} as the output of a single run of
\code{generate_topline_multiselect()}, \emph{if it suceeds.} The \code{error} list captures all runs of
\code{generate_topline_multiselect()} that have failed, returning the error messages. If all run fails,
\code{result} will be a \code{NULL} list; on the other end, if no run fails, \code{error} will be a \code{NULL} list.
}
\description{
This function is the second part of the \strong{split-apply-combine} workflow in handling multiple selection questions
in the survey. It takes the result of \code{split_df()}. The \code{caption} argument can also be generated with the helper function.
}
\examples{
\donttest{
# Vector of patterns
patterns <- c("prefix_1", "predix_2", ...)

# List of data.tables
list_df <- split_df(df, patterns, "weight_var")

# Apply topline generation to each element of 'list_df'
captions <- list(c("caption_parent", "caption_child"), "caption_2", ...)
parents <- c(TRUE, FALSE, ...)
results <- apply_topline_multiselect(list_df, "weight_var", captions, parents)
}
}
\seealso{
\code{\link[=generate_topline_multiselect]{generate_topline_multiselect()}} for single topline generation for multiple selection questions,
\code{\link[=split_df]{split_df()}} for splitting a single data frame into smaller subsets, and \code{\link[=combine_topline_multiselect]{combine_topline_multiselect()}}
for combining the multiselect toplines.
}
