% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{calc_depth}
\alias{calc_depth}
\title{Calculate depth from pressure in SEAWATER (factors-in gravity and latitude).}
\usage{
calc_depth(latitude, pressure)
}
\arguments{
\item{latitude}{Latitude in decimal degrees.}

\item{pressure}{Pressure in decibars.}
}
\value{
A data frame containing depth in meters from pressure and adjusted for latitude.
}
\description{
Calculate depth from pressure in SEAWATER (factors-in gravity and latitude).
}
