% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_patterns}
\alias{check_patterns}
\title{Check if column names start with patterns}
\usage{
check_patterns(mainstrs, patterns)
}
\arguments{
\item{mainstrs}{A character vector of strings that are checked against each element in \code{patterns}.}

\item{patterns}{A character vector of pattern strings.}
}
\value{
A logical vector with the same length as \code{patterns}, indicating whether each pattern is one that
at least one string in \code{mainstrs} starts with.
}
\description{
This helper function takes two character vectors \code{mainstrs} and \code{patterns}, returning a boolean indicating
whether each element of \code{patterns} is a string pattern in which at least one element of \code{mainstrs} starts with.
In other words, for each element of \code{patterns}, does any element of \code{mainstrs} start with this particular pattern?
If so, return \code{TRUE}. This helper can be used to efficiently verify the argument \code{patterns} in \code{split_df()},
especially when there is a large number of patterns.
}
\examples{
\donttest{
# Vector of patterns
patterns <- c("prefix_1", "predix_2", ...)

# Checking if these patterns are valid for 'df'
check_patterns(mainstrs = names(df), patterns) |> all()

# Find all invalid patterns
valid_index <- check_patterns(mainstrs = names(df), patterns)
patterns[!valid_index]
}
}
