% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_checks.R
\name{chk_numeric_no_na}
\alias{chk_numeric_no_na}
\title{Check for Numeric Vector without Missing Values}
\usage{
chk_numeric_no_na(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{A vector to check.}

\item{arg_name}{Internal argument name for use in error messages.}
}
\description{
Internal function to validate if an input vector is numeric and
contains no missing values. Stops execution if validation fails.
}
\keyword{internal}
