% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{classify_topics}
\alias{classify_topics}
\title{classify_topics()}
\usage{
classify_topics(dataframe, column, output = "", topic_aliases = "",
  confidence = FALSE)
}
\arguments{
\item{dataframe}{dataframe or tibble of survey responses}

\item{column}{string variable of free text responses to which the determine_topics() function
has already been applied in this session}

\item{output}{name of the new column to be produced, defaults to '{column}_topic'}

\item{topic_aliases}{named string vector, denoting for each topic (1,2.. etc.) what it is to
be renamed. Leave blank to stay as 'topic1', 'topic2' etc.}

\item{confidence}{logical indicating whether to include topic confidence in output (how likely
that the classified topic is definitive)}
}
\value{
the original dataframe with an additional column (name specified by output) containing which
topic each response in 'column' falls in to, optionally aliased
}
\description{
Function which takes in a dataframe and column and produces another column,
denoting which topic each response is most relevant to
}
