% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_column.R
\name{clean_at}
\alias{clean_at}
\title{clean_at}
\usage{
clean_at(data, ..., num = FALSE, null_response = TRUE)
}
\arguments{
\item{data}{dataframe or tibble with a row per survey response}

\item{...}{columns for the clean() function to be applied to}

\item{num}{logical indicating whether numbers should be removed as part of cleaning process}

\item{null_response}{logical indicating whether to remove non-answers (e.g. 'no', 'none', 'na' etc.)}
}
\value{
Original dataframe with specified columns cleaned and standardised ready for analysing
}
\description{
Apply clean() function to multiple specified columns
}
\seealso{
Other preprocess functions: \code{\link{clean_all}},
  \code{\link{clean}}
}
\concept{preprocess functions}
