% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_column.R
\name{clean_column}
\alias{clean_column}
\title{Function to clean single text column, remove punctuation and standardise}
\usage{
clean_column(data, column, num = TRUE, null_response = TRUE)
}
\arguments{
\item{data}{dataframe or tibble with a row per survey response}

\item{column}{name of a character column in the data frame to be cleaned}

\item{num}{logical indicating whether numbers should be removed as part of cleaning process}
}
\value{
Original dataframe with text column cleaned and standardised ready for analysing
}
\description{
Function to clean single text column, remove punctuation and standardise
}
