% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clt_demo.R
\name{clt_demo}
\alias{clt_demo}
\title{Central Limit Theorem Demo}
\usage{
clt_demo(n = 20, N = 10000, distribution = "binormal", a = 30, b = 50, s = 6)
}
\arguments{
\item{n}{The sample size to draw from a specified distribution. This defaults to n=30,
but can be experimented with to see what the distribution looks like when
under larger and smaller sample sizes. This method requires n >= 2.}

\item{N}{The total number of replications to make in the simulation. This number is
saying how many data points to use when constructing the histogram, so it should
be fairly large. This method must be N >= 100.}

\item{distribution}{One of "binormal", "uniform", or "gamma".}

\item{a}{In each distribution, this has a different function.
\itemize{
\item binormal The mean of one of the normal distributions
\item uniform  The left endpoint of a uniform distribution
\item gamma    The shape parameter in a gamma distribution
}}

\item{b}{As with a, b specifies a specific parameter in a distribution. Note in the
binormal case, a = b will produce a unimodal normal distribution.
\itemize{
\item binormal The mean of one of the normal distributions
\item uniform  The right endpoint of a uniform distribution
\item gamma    The rate parameter of a uniform distribution
}}

\item{s}{The standard deviation for a normal distribution. Can be ignored or specified
as NULL when using other distributions, although this isn't strictly necessary.}
}
\value{
A list with
\itemize{
\item data Samples drawn from the specified distribution
\item plot A histogram of the results
\item popmean The population mean for the specified distribution
\item popvariance The population variance for the specified distribution
}
}
\description{
Provides a flexible demonstration of the central limit theorem. Particularly,
this function uses the theoretical parameters from one of three specific
distributions to show the distribution of the sample mean approaches a normal
distribution. The user can choose between three distributions:
\itemize{
\item binormal A binormal distribution \eqn{X|Z ~ N(Z, s)}
\item uniform A uniform distribution such that \eqn{X ~ U[a, b]}
\item gamma A gamma distribution, \eqn{X ~ \Gamma(a, b)}
}

The central limit theorem essentially guarantees that, for a large enough
sample size of independent samples, the distribution of the mean from a
probability-based sample converges to a standard normal distribution.
This is the basis for using normal and t-distributed confidence intervals.
}
\examples{
# Can be run without passing any arguments
clt <- clt_demo()
# Use clt$plot to obtain the graph
}
\references{
Billingsley, Patrick. Probability and measure. John Wiley & Sons, 2008.
}
