% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearning-helpers.R
\name{col_nms_rm_splchar}
\alias{col_nms_rm_splchar}
\title{Remove special characters column names}
\usage{
col_nms_rm_splchar(df)
}
\arguments{
\item{df}{A data frame or tibble.}
}
\value{
A data frame or tibble.
}
\description{
Remove special characters from column names in a data frame. To be precise, the function
is implemented such that only the character class of numbers and letters in the current locale,
i.e., \verb{[:alnum:]} is retained.
}
\examples{
# Remove special characters from column names
new_df <- col_nms_rm_splchar(df)
}
\seealso{
\code{\link[=rm_spl_char]{rm_spl_char()}} for removing special characters from selected columns.
}
