% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearning-helpers.R
\name{col_nms_to_title}
\alias{col_nms_to_title}
\title{Capitalize column names}
\usage{
col_nms_to_title(df)
}
\arguments{
\item{df}{A data frame or tibble.}
}
\value{
A data frame or tibble.
}
\description{
Convert column names in a data frame to title cases.
}
\examples{
# Capitalize column names
new_df <- col_nms_to_title(df)
}
