% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_topline_multiselect.R
\name{combine_topline_multiselect}
\alias{combine_topline_multiselect}
\title{Combine the results of \code{apply_topline_multiselect()}}
\usage{
combine_topline_multiselect(l)
}
\arguments{
\item{l}{A list of lists as returned by \code{apply_topline_multiselect()}.}
}
\value{
A list of \code{flextable} objects.
}
\description{
This function completes the \strong{split-apply-combine} workflow in handling multiple selection questions
in the survey. This is a helper function that accepts the output of \code{apply_topline_multiselect()} containing two
lists--- \code{return} and \code{error}. It will either \strong{1)} throw an error indicating that attention must be paid to
the \code{error} list, or \strong{2)} return a single list of toplines, which can then be passed to \code{print_tbls()}.
}
\examples{
\donttest{
# Vector of patterns
patterns <- c("prefix_1", "predix_2", ...)

# List of data.tables
list_df <- split_df(df, patterns, "weight_var")

# Apply topline generation to each element of 'list_df'
captions <- list(c("caption_parent", "caption_child"), "caption_2", ...)
parents <- c(TRUE, FALSE, ...)
results <- apply_topline_multiselect(list_df, "weight_var", captions, parents)

# Combine the results
list_of_toplines <- combine_topline_multiselect(l = results)
}
}
