% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{conf_int}
\alias{conf_int}
\title{conf_int}
\usage{
conf_int(.tbl, .alpha, .interval = "t")
}
\arguments{
\item{.tbl}{Output from \code{mk_stat} or \code{stratified_estimator}}

\item{.alpha}{The confidence level.}

\item{.interval}{Either "t" or "z". If "t", degrees of freedom are inferred from the input .tbl}
}
\value{
A list with
\item{level}{The percentage of the confidence interval, 100(1-alpha)}
\item{df}{\code{NA} if .interval="z", otherwise the degrees of freedom used in t interval}
\item{lower}{The lower bound of the confidence interval}
\item{upper}{The upper bound of the confidence interval}
}
\description{
A helper function for generating symmetric confidence intervals using a \code{t}
or standard normal distribution. The confidence interval is a
100(1 - alpha)\% confidence interval.
}
\examples{
# Simplified output similar to stratified_sample
ds <- tibble::tibble(Ntotal=10, ntotal=4, point=3, se=1)
conf_int(ds, 0.05, "t")

# Simplified output from mk_stat
d <- tibble::tibble(n = 4, point=3, se=1)
conf_int(d, 0.05, "t")
}
