% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{create_ev_gps}
\alias{create_ev_gps}
\title{Create gps.csv files used by Echoview.}
\usage{
create_ev_gps(df, lat = df$lat, long = df$long, date_time = df$datetime)
}
\arguments{
\item{lat}{Latitude in decimal degrees.}

\item{long}{Longitude in decimal degrees.}

\item{date_time}{Date/time field in ISO 8601 date format (e.g., %Y-%m-%d %H:%M:%S).}
}
\value{
A data frame containing  \code{Date}, \code{Time}, \code{latitude}, and \code{longitude}.
}
\description{
Create gps.csv files used by Echoview.
}
