% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{dd2decmin}
\alias{dd2decmin}
\title{Convert decimal degrees to degrees and decimal minutes}
\usage{
dd2decmin(dd, format = "latex")
}
\arguments{
\item{dd}{Latitude or longitude in decimal degrees.}
}
\value{
Latitude or longitude in degrees and decimal minutes.
}
\description{
Convert decimal degrees to degrees and decimal minutes
}
