% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{determine_topics}
\alias{determine_topics}
\title{determine_topics()}
\usage{
determine_topics(dataframe, column, num_topics = 2)
}
\arguments{
\item{dataframe}{dataframe or tibble of survey responses}

\item{column}{string variable of free text responses to which the LDA model will be applied}

\item{num_topics}{integer denoting the number of topics (k) to be specified for the LDA model, defaults to 2}
}
\value{
a dataframe, where each column is a topic and contains the 5 most common words in those topics
}
\description{
Function which takes in a dataframe and column and abstracts the use of
the topicmodels package by applying an LDA model and saving data internally
to be used later by other *_topics functions
}
