% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{ev2posix}
\alias{ev2posix}
\title{Convert time from Echoview format to POSIXct}
\usage{
ev2posix(date, time, tz = "GMT")
}
\arguments{
\item{date}{Date in Echoview format.}

\item{time}{Time in Echoview format.}
}
\value{
A date/time object in POSIXct format.
}
\description{
Convert time from Echoview format to POSIXct
}
