% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning-class.R
\name{FrameCleaner}
\alias{FrameCleaner}
\title{A class for cleaning a data frame object}
\description{
This class contains common cleaning routines for data frame objects. The input data frame
may be one of \code{data.frame}, \code{tibble}, or \code{data.table}, but the output data frame will be
a \code{data.table} object. If needed, an object of the original class can be returned by
calling the \code{restore()} method, which returns a copy of the modified data frame with
the original \code{class} attribute.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{df}}{An object that inherits from \code{data.frame}.}

\item{\code{col_nms}}{Column names of the input \code{df}.}

\item{\code{ptype}}{A 0-observation slice of the input \code{df}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FrameCleaner$new()}}
\item \href{#method-col_nms_trim}{\code{FrameCleaner$col_nms_trim()}}
\item \href{#method-restore}{\code{FrameCleaner$restore()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a new \code{FrameCleaner} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameCleaner$new(df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{An object that inherits from \code{data.frame}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{FrameCleaner} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-col_nms_trim"></a>}}
\if{latex}{\out{\hypertarget{method-col_nms_trim}{}}}
\subsection{Method \code{col_nms_trim()}}{
A method for removing leading and trailing white spaces, including
space (\code{' '}), form feed (\verb{\\f}), line feed (\verb{\\n}), carriage return (\verb{\\r'},
horizontal tab (\verb{\\t}), and vertical tab (\verb{\\v}) from column names.
The function modifies the \code{df} field in-place.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameCleaner$col_nms_trim()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-restore"></a>}}
\if{latex}{\out{\hypertarget{method-restore}{}}}
\subsection{Method \code{restore()}}{
A method for restoring the original class attribute of the input data frame.
Note that a copy is returned if the original input \code{df} is not a \code{data.table}.
If the original input \code{df} is a \code{data.table}, then this method should be a
no-op.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameCleaner$restore()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An object that inherits from \code{data.frame}.
}
}
}
