% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_generate_tbls.R
\name{ft_generate_tbls}
\alias{ft_generate_tbls}
\title{Iteratively generate tables for 'Feeling Thermometer' variables}
\usage{
ft_generate_tbls(l, df, weight, type = "topline", output = "word")
}
\arguments{
\item{l}{A list of arguments to be passed on to either \verb{generate_xtab_*()} or \verb{generate_topline_*()}:
\describe{
\item{For toplines}{
\preformatted{list_topline <- tibble::tribble(
     ~x, ~caption,
     "var_1", "caption1",
     "var_2", "caption2",
     "var_3", "caption3",
     .
     .
     .
  )}}
\item{For crosstabs}{
\preformatted{list_xtab <- tibble::tribble(
     ~x, ~y, ~caption,
     "var_1", "var_2" "caption1",
     "var_2", "var_9", caption2",
     "var_3", "var_23", "caption3",
     .
     .
     .
  )}}
}}

\item{df}{A data frame or tibble.}

\item{weight}{A length one character vector used as the caption for the topline.}

\item{type}{Must either be 'topline' or 'crosstab'. Defaults to 'topline'.}

\item{output}{Must either be 'latex' or 'word'. Defaults to 'word'.}
}
\value{
A list containing elements that are length one character vectors (latex) or lists objects (word), each of which is the source code for a single
table (latex) or a \code{s3} list object (word).
}
\description{
This function provides a wrapper around \verb{ft_generate_xtab_*()} and \verb{ft_generate_topline_*()} (all internal),
allowing users to efficiently generate many tables by passing a list of arguments. The shapes of the argument lists
may vary depending on the table \code{type}.
}
\examples{
\donttest{
# Generate crosstabs
list_of_xtabs <- ft_generate_tbls(
  l = list_xtab,
  df,
  "weight",
  "crosstab",
  "word"
)

# Generate toplines
list_of_topline <- ft_generate_tbls(
  l = list_topline,
  df,
  "weight",
  "topline",
  "latex"
)
}
}
\seealso{
\code{\link[=generate_xtab_docx]{generate_xtab_docx()}}, \code{\link[=generate_xtab_latex]{generate_xtab_latex()}}, \code{\link[=generate_topline_docx]{generate_topline_docx()}}, and \code{\link[=generate_topline_latex]{generate_topline_latex()}} for single crosstab or topline generation.
}
