% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tbls.R
\name{generate_tbls}
\alias{generate_tbls}
\title{Iteratively generate multiple tables}
\usage{
generate_tbls(l, df, weight, type = "topline")
}
\arguments{
\item{l}{A list of arguments to be passed on to either \code{generate_xtab()}, \code{generate_xtab_3way()}, or \code{generate_topline()}:
\describe{
\item{For toplines}{
\preformatted{list_topline <- tibble::tribble(
     ~x, ~caption,
     "var_1", "caption1",
     "var_2", "caption2",
     "var_3", "caption3",
     .
     .
     .
  )}}
\item{For crosstabs}{
\preformatted{list_xtab <- tibble::tribble(
     ~x, ~y, ~caption,
     "var_1", "var_2", "caption1",
     "var_2", "var_9", "caption2",
     "var_3", "var_23", "caption3",
     .
     .
     .
  )}}
\item{For three-way crosstabs}{
\preformatted{list_xtab <- tibble::tribble(
     ~x, ~y, ~z, ~caption,
     "var_1", "var_2", "var_3", "caption1",
     "var_2", "var_9", "var_1", "caption2",
     "var_3", "var_23", "var_17", "caption3",
     .
     .
     .
  )}}
}}

\item{df}{A data frame or tibble.}

\item{weight}{A length one character vector used as the caption for the topline.}

\item{type}{Must either be 'topline', 'crosstab_2way', or 'crosstab_3way'. Defaults to 'topline'.}
}
\value{
A list containing elements that are list objects, each of which is the source code for a single
\code{s3} list object of class \code{flextable}.
}
\description{
This function provides a wrapper around \code{generate_xtab()}, \code{generate_xtab_3way()}, and \code{generate_topline()},
allowing users to efficiently generate many tables by passing a list of arguments. The shapes of the argument
lists may vary depending on the table \code{type}.
}
\examples{
\donttest{
# Generate crosstabs
list_of_xtabs <- generate_tbls(
  l = list_xtab,
  df,
  "weight",
  "crosstab_2way"
)

#' list_of_xtabs <- generate_tbls(
  l = list_xtab,
  df,
  "weight",
  "crosstab_3way"
)

# Generate toplines
list_of_topline <- generate_tbls(
  l = list_topline,
  df,
  "weight",
  "topline"
)
}
}
\seealso{
\code{\link[=generate_xtab]{generate_xtab()}}, \code{\link[=generate_xtab_3way]{generate_xtab_3way()}}, and \code{\link[=generate_topline]{generate_topline()}} for single crosstab or topline generations.
}
