% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_topline.R
\name{generate_topline_latex}
\alias{generate_topline_latex}
\title{Generate toplines (latex)}
\usage{
generate_topline_latex(df, x, weight, caption)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{x}{A single string of variable name.}

\item{weight}{A single string of the weighting variable.}

\item{caption}{A length one character vector used as the caption for the topline.}
}
\value{
A character vector of the table source code, which is an object of class \code{knitr_kable}.
}
\description{
This function takes a data frame or a tibble object, a variable of interest, and a string
caption, returning a length one character vector of the table source code. The function
implements data masking internally so users must specify data variables as strings
(i.e. use "x", "weight").
}
\examples{
\donttest{
# Generate a topline
df \%>\% generate_topline_latex(df = ., "x_var", "weight", "X")
}
}
\seealso{
\code{\link[=generate_topline_docx]{generate_topline_docx()}} for outputting to Microsoft word.
}
