% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_topline_multiselect.R
\name{generate_topline_multiselect}
\alias{generate_topline_multiselect}
\title{Generate toplines for multiple selection questions (word)}
\usage{
generate_topline_multiselect(df, weight, caption, parent = FALSE)
}
\arguments{
\item{df}{A data frame or data.table or tibble.}

\item{weight}{A single string of the weighting variable.}

\item{caption}{A length one character vector used as the caption for the topline.}

\item{parent}{A boolean indicating whether \code{df} has a parent response column. Defaults to \code{FALSE}.}
}
\value{
A list object containing a either one or two objects of class \code{flextable}.
}
\description{
This function generates a \emph{single} topline for cases where there is no \strong{parent} response column in the
multiselect question. These cases are organized such that each column in \code{df} represents a single choice
in a multiselect question. Alternatively, there may be cases where \code{df} is structured such that the
\strong{first} column represents a \strong{parent} of the multiselect question in addition to the \strong{choices}
or \strong{children} columns that represent the multiselect choices. In such cases, \emph{two} toplines will
be generated--- one for the parent response and another for the choices columns. Users may control the output
with the \code{parent} argument. Note that, for multiselect questions that involve a parent response column,
it is assumed that such a column is the first column of the argument \code{df}. Therefore, some pre-processing may be
needed to ensure that this requirement is met.
}
\examples{
\donttest{
# Generate a topline (multiselect)
df \%>\%
  generate_topline_multiselect(
    df = .,
    "weight",
    "caption",
    FALSE
  )
}
}
\seealso{
\code{\link[=apply_topline_multiselect]{apply_topline_multiselect()}} for vectorized topline generation for multiple selection questions.
}
