% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xtab.R
\name{generate_xtab}
\alias{generate_xtab}
\title{Generate crosstabs}
\usage{
generate_xtab(df, x, y, weight, caption)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{x}{A single string of the independent variable.}

\item{y}{A single string of the dependent variable.}

\item{weight}{A single string of the weighting variable.}

\item{caption}{A length one character vector used as the caption for the crosstab.}
}
\value{
A character vector of the table source code, which is an object of class \code{knitr_kable}.
}
\description{
This function takes a data frame or a tibble object, a few variables, and a string caption, returning
a length one character vector of the table source code. The function implements data masking internally
so users must specify data variables as strings (i.e. use "x", "y", "weight").
}
\examples{
\donttest{
# Generate a crosstab
df \%>\% generate_xtab(df = ., "x_var", "y_var", "weight", "X by Y")
}
}
\seealso{
\code{\link[=generate_topline]{generate_topline()}} for topline generation.
}
