% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_arguments.R
\name{generate_xtab_3way_args}
\alias{generate_xtab_3way_args}
\title{Generate a list of three-way crosstab arguments}
\usage{
generate_xtab_3way_args(df, control_var, independent_vars, dependent_vars)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{control_var}{A single string of control variable.}

\item{independent_vars}{A character vector of independent variables.}

\item{dependent_vars}{A character vector of dependent variables.}
}
\value{
A list of arguments. If n tables were to be generated, this is a \code{tibble} with dimensions (n, 4).
}
\description{
This function is helper for \code{generate_tbls()} or \code{ft_generate_tbls()}. It automatically generates a list of arguments to be
passed on to \code{generate_tbls()} or \code{ft_generate_tbls()} for a given set of arguments. This may be useful when the number of
variables in a data frame is large, and there is a need to generate varying arguments efficiently.
}
\examples{
\donttest{
# Generate list of arguments
generate_xtab_3way_args(df, "control_var", c("independent_1", ...), c("dependent_1", ...))
}
}
\seealso{
\code{\link[=generate_tbls]{generate_tbls()}} or \link{ft_generate_tbls} for an example of a list of arguments.
}
