% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xtab.R
\name{generate_xtab_docx}
\alias{generate_xtab_docx}
\title{Generate crosstabs (word)}
\usage{
generate_xtab_docx(df, x, y, weight, caption)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{x}{A single string of the independent variable.}

\item{y}{A single string of the dependent variable.}

\item{weight}{A single string of the weighting variable.}

\item{caption}{A length one character vector used as the caption for the crosstab.}
}
\value{
A list object, which is an object of class \code{flextable}.
}
\description{
This function takes a data frame or a tibble object, a few variables, and a string caption, returning
a list object that is an instance of class \code{flextable}. The function implements data masking internally
so users must specify data variables as strings (i.e. use "x", "y", "weight").
}
\examples{
\donttest{
# Generate a crosstab
df \%>\% generate_xtab_docx(df = ., "x_var", "y_var", "weight", "X by Y")
}
}
\seealso{
\code{\link[=generate_xtab_latex]{generate_xtab_latex()}} for outputting to pdf.
}
