% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_xtab_args.R
\name{list_xtab_args}
\alias{list_xtab_args}
\title{Generate multiple lists of crosstab arguments}
\usage{
list_xtab_args(df, var_of_interest, dependent_vars = NULL, rm = NULL)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{var_of_interest}{A character vector of variables of interest.}

\item{dependent_vars}{A list of character vectors, each of which contains dependent variables. Defaults
to crossing each element of \code{var_of_interest} with all other variables in \code{df} if \code{rm} is \code{NULL}.}

\item{rm}{A character vector of variables to be removed from consideration. Defaults to \code{NULL}.}
}
\value{
A list containing two elements--- \code{result} and \code{error}, which are lists in and of themselves.
These two lists have the same structure and number of elements. The \code{result} is a \strong{named} list
of \code{tibble} objects. One can think of each \code{tibble} as the output of a single run of \code{generate_xtab_args()},
\emph{if it suceeds.} The \code{error} list captures all runs of \code{generate_xtab_args()} that have failed,
returning the error messages. If all run fails, \code{result} will be a \code{NULL} list; on the other end, if no run
fails, \code{error} will be a \code{NULL} list. Users can easily find out which \code{var_of_interest} failed by examining
the \code{error} list and running the following code (assuming that the list is called \code{list_of_args} and the
vector of variables of interest is called \code{var_of_interest}):
\preformatted{
        lgl_index <- purrr::map_lgl(list_of_args[["error"]], is.null)
        var_of_interest[!lgl_index]
  }
To examine the results of the successful runs, use:
\preformatted{
        list_of_args[["result"]][lgl_index]
  }
Or, if all runs are successful, simply return the \code{result} list:
\preformatted{
        list_of_xtab_args <- list_of_args[["result"]]
  }
}
\description{
Simply put, \code{list_xtab_args()} is a vectorized version of \code{generate_xtab_args()}. The function
\code{generate_xtab_args()} allows for one \code{var_of_interest} at a time. There may be use cases where
multiple \code{var_of_interest} must be considered. In those cases, this function may be useful as it
allows users to pass a vector of \code{var_of_interest} and a corresponding list of \code{dependent_vars}.
These two vectors must be of the same length in terms of \code{vec_size()}. One may think of these two
vectors as pairs in the sence that--- each ith element in \code{var_of_interest} is crossed with each
element of \code{dependent_vars}, which is a vector in and of itself. If no \code{dependent_vars} list is
specified, the function defaults to crossing each variable in \code{var_of_interest} with all other
columns in \code{df} unless \code{rm} is specified.
}
\examples{
# Create var_of_interest vector
var_of_interest <- c("col1", "col3", "col20", "col23", "col32")

# List of dependent variable vectors
dependent_vars <- list(
  c("col2", "col4", "col5", "col100"),
  c("col2", "col4"),
  c("col2", "col4", "col5", "col86"),
  NULL,
  c("col2", "col4"),
)

# Create list_of_args
list_of_args <- list_xtabs_args(df_, var_of_interest, dependent_vars)
}
