% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{load_pkgs}
\alias{load_pkgs}
\title{Install and load all packages provided from a character vector}
\usage{
load_pkgs(pkgs)
}
\arguments{
\item{pkgs}{A character vector containing the names of packages to be loaded and/or installed.}
}
\description{
Install and load all packages provided from a character vector
}
