% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary.R
\name{make_summary}
\alias{make_summary}
\title{Summarize data from a stratified random sample}
\usage{
make_summary(
  .data,
  .group,
  .y,
  .group_N,
  .stat = "mean",
  .group_weights = NULL,
  .fpc = TRUE
)
}
\arguments{
\item{.data}{A dataframe or tibble of data}

\item{.group}{A stratifying variable as a column in the data}

\item{.y}{An observation variable in the data}

\item{.group_N}{An exogeneous vector of population sizes for each subgroup to join to the
data when computing stratified effects.}

\item{.stat}{One of "mean", "total" or "prop"}

\item{.group_weights}{Optional column of weights in the dataframe.
If weights are given in the dataframe, they will be used in place of the
population vector (.group_N), although .group_N is still a required argument.}

\item{.fpc}{Logical variable: indicates whether to compute a finite popoulation correction.}
}
\value{
A tibble (\code{data.frame}-like structure)
}
\description{
Support for computing tables of statistics for stratified variables.
Wraps \code{mk_stat} for each subgroup (stratum).
Only supports stratification by one variable at this time.
}
\examples{
d <- data.frame("stratum"=rep(1:5, 6), "y"=rnorm(30, 30, 3))
N <- 50
make_summary(d, stratum, y, N)
}
