% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs_functions.R
\name{mk_stat}
\alias{mk_stat}
\title{Estimate a statistic from a simple random sample design}
\usage{
mk_stat(y, stat = "mean", N = NULL, fpc = TRUE, weights = NULL)
}
\arguments{
\item{y}{A numeric vector with at least two elements}

\item{stat}{A string, either "mean", "total", or "prop" (proportion)}

\item{N}{The population size.
This must be supplied and must be larger than length(y).
If unknown, recommend passing it as some N > n with fpc=FALSE.}

\item{fpc}{TRUE/FALSE.
Whether or not to compute a finite population correction (for variance).}

\item{weights}{Optional.
If passed, it must be a vector of weights at least as long as the input y.}
}
\value{
tibble with
\itemize{
\item n (sample size)
\item point estimate (stat: mean, total, proportion)
\item estimator variance (not the same as population variance)
\item estimator standard error
\item coefficient of variation
}
}
\description{
Function to get point estimates and statistics associated with them.
Given a point estimate (mean, total, prop), \code{mk_stat} will generate the
\itemize{
\item sample size, n, at calling time
\item point estimate
\item estimator variance
\item standard error
\item coefficient of variation
assuming a simple random sample of data.
}
}
\examples{
N <- 10
n <- 5
y <- c(0, 0, 0, 1, 1)
weights <- rep(N/n, n)
mk_stat(y, stat="mean", N, fpc=TRUE, weights=weights)
# A tibble: 1 x 5
#      n point   var    se    cv
#  <int> <dbl> <dbl> <dbl> <dbl>
#1     5   0.4  0.15 0.387 0.968

N <- 10
n <- 5
y <- c(0, 0, 0, 1, 1)
weights <- rep(N/n, n)
mk_stat(y, stat="total", N, fpc=TRUE, weights=weights)
# A tibble: 1 x 5
#      n point   var    se    cv
#  <int> <dbl> <dbl> <dbl> <dbl>
#1     5     4    15  3.87 0.968

#' N <- 10
n <- 5
y <- c(0, 0, 0, 1, 1)
weights <- rep(N/n, n)
mk_stat(y, stat="prop", N, fpc=TRUE, weights=weights)
# A tibble: 1 x 5
#      n point   var    se    cv
#  <int> <dbl> <dbl> <dbl> <dbl>
#1     5   0.4  0.03 0.173 0.433
}
\references{
Lohr, Sharon L. Sampling: Design and Analysis. Chapman and Hall/CRC, 2019.
}
