% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_grams.R
\name{n_grams}
\alias{n_grams}
\title{Function to tabulate a text column, showing the most commonly used words, optionally broken down by another column(s)}
\usage{
n_grams(data, column, word = "", n = 2, min = 2)
}
\arguments{
\item{data}{dataframe or tibble with a row per survey response}

\item{column}{name of a character column in the data frame to be tabulated}

\item{word}{optional string indicating how to filter the resulting dataframe (only return ngrams containing 'word')}

\item{n}{number indicating how many most common words to return for each group. Defaults to 3}

\item{min}{number indicating the minimum number of times a word needs to appear for it to be included in output, defaults to 5}
}
\value{
Original dataframe with text column cleaned and standardised ready for analysing
}
\description{
Function to tabulate a text column, showing the most commonly used words, optionally broken down by another column(s)
}
