% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{num2words}
\alias{num2words}
\title{Convert numbers less than 9 to words.}
\usage{
num2words(x)
}
\arguments{
\item{x}{Number to convert to letters}
}
\value{
If greater than 10, a number, else the number spelled-out in words.
}
\description{
Convert numbers less than 9 to words.
}
