% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{numbers2words}
\alias{numbers2words}
\title{Convert numbers greater than 9 to words.}
\usage{
numbers2words(x)
}
\arguments{
\item{x}{Number to convert to letters}
}
\value{
If less than 10, a number, else the number spelled-out in words.
}
\description{
Convert numbers greater than 9 to words.
}
