% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print_tbls}
\alias{print_tbls}
\title{Print tables iteratively}
\usage{
print_tbls(l)
}
\arguments{
\item{l}{A list of \code{flextable} objects, which are s3 lists with certain structures.}
}
\value{
Invisible.
}
\description{
This function takes a list of \code{flextable} objects and prints the raw code. \strong{Note} that chunk option
results must be set to \strong{asis}, i.e., \code{results='asis'}.
}
\examples{
\donttest{
# Toplines
print_tbls(list_toplines)

# Crosstabs
print_tbls(list_xtabs)
}
}
