% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{process_check_all_that_apply_q}
\alias{process_check_all_that_apply_q}
\title{process_check_all_that_apply_q: Process a 'check all answers that apply' question}
\usage{
process_check_all_that_apply_q(mytable, question)
}
\arguments{
\item{mytable}{the dataset}

\item{question}{question under analysis (use the janitor::clean_names() text)}
}
\value{
a tibble: options (rows) x count
}
\description{
process_check_all_that_apply_q: Process a 'check all answers that apply' question
}
