% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{process_check_all_that_apply_q_by_grp}
\alias{process_check_all_that_apply_q_by_grp}
\title{Process a check-all-that-apply question, but break down numbers by group}
\usage{
process_check_all_that_apply_q_by_grp(mytable, mygroup, question)
}
\arguments{
\item{mytable}{the dataset}

\item{mygroup}{column whose values defines a group}

\item{question}{question under analysis (use the janitor::clean_names() text)}
}
\value{
a tibble of questions (cols) x grps (rows)
}
\description{
Process a check-all-that-apply question, but break down numbers by group
}
