% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex-helpers.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{usepackage_latex}
\alias{kableExtra_latex_packages}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{kableExtra}{\code{\link[kableExtra]{kableExtra_latex_packages}}, \code{\link[kableExtra]{usepackage_latex}}}
}}

