% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearning-helpers.R
\name{rm_spl_char}
\alias{rm_spl_char}
\title{Remove special characters from selected columns}
\usage{
rm_spl_char(df, var)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{var}{A character vector, with length no greater than \code{length(df)}, specifying the columns
from which special characters are to be removed.}
}
\value{
An object of the same type as \code{df} with specified columns stripped of their special characters.
}
\description{
This function takes a data frame or a tibble object and a vector of columns, removing special
characters from the specified columns in the data frame or tibble. To be precise, the function
is implemented such that only the character class of numbers and letters in the current locale,
i.e., \verb{[:alnum:]}, and periods (".") are retained.
}
\examples{
\dontrun{
# Remove special characters from columns 1 & 2
df_ \%>\% rm_spl_char(var = c("col_1", "col_2"))

# Remove special characters from all columns
cols <- names(df_)
df_ \%>\% rrm_spl_char(var = cols)
}
}
\seealso{
\code{\link[=col_nms_rm_spl_char]{col_nms_rm_spl_char()}} for removing special characters from columns names.
}
