% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{s_plot}
\alias{s_plot}
\title{Simple plotting tools}
\usage{
s_plot(.data, .y, .group = NULL, kind = NULL)
}
\arguments{
\item{.data}{A tibble dataset.}

\item{.y}{A column of numeric observations in .data}

\item{.group}{Optional. A column of population subgroups in .data}

\item{kind}{Optional. Either "hist" or "histogram". This argument should only be passed
when a .group is passed in the function. It will throw an error when .group
is not specified with it.}
}
\value{
A ggplot object (graph)
}
\description{
A simple interface for plotting data using \code{ggplot2} as a backend.
Provides support for plotting histograms of simple random samples or colorized
boxplots for stratified samples. Optionally, the plot for a stratified sample
can be changed to a colorized histogram. Much greater flexibility is available
through \code{ggplot2}, but this is a helpful starter tool for using that library.

In histograms, bins are automatically computed using Doane's formula \link{1}. This
is an immutable default choice for this implementation. Using \code{ggplot}
directly will provide greater flexibility for specifying bin widths and other
options.
}
\examples{
d <- tibble::tibble(s = rep(1:3, 30), y = rnorm(90, 30*s, 5*sqrt(s)))
s_plot(d, y, s, kind = "histogram") + ggplot2::labs(
title = "My cool plot",
subtitle = "These can be entirely customized as ever you'd like",
caption = "See more from ggplot help files",
x = "An observed variable (with what units)",
y = "Frequencies",
legend = "Some group variable"
) + ggplot2::theme_bw()
}
\references{
\link{1} \cite{https://en.wikipedia.org/wiki/Histogram#Doane's_formula}
}
