% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{scs2dd}
\alias{scs2dd}
\title{convert latitude or longitude from SCS format to decimal degrees}
\usage{
scs2dd(x)
}
\arguments{
\item{x}{Latitude or longitude in SCS format.}
}
\value{
Latitude or longitude in decimal degrees.
}
\description{
convert latitude or longitude from SCS format to decimal degrees
}
