% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_score.R
\name{sentiment_score}
\alias{sentiment_score}
\title{sentiment_score}
\usage{
sentiment_score(data, column, output = "", display = "simple")
}
\arguments{
\item{data}{dataframe or tibble with a row per survey response}

\item{column}{name of a character column in the data frame which can be used to calculate the sentiment}

\item{output}{string representing the name of the new sentiment column, defaults to '[column]_sentiment'}

\item{display}{string denoting whether the sentiment should be represented by: words denoting the sentiment ('simple' - the default) the actual score ('num') or
the sentiment of every response should be ranked with a higher rank denoting more positive sentiment i.e. 1 being the most positive ('rank')}
}
\value{
Original dataframe with an additional column denoting the sentiment of the response
}
\description{
Return a sentiment score for each response, in a variety of different ways
}
