% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_sample.R
\name{stratified_stat}
\alias{stratified_stat}
\title{stratified_stat}
\usage{
stratified_stat(.tbl, .stat = "mean", .fpc = TRUE)
}
\arguments{
\item{.tbl}{Output from (or similar to) \code{make_summary}}

\item{.stat}{One of "mean", "total", or "prop"}

\item{.fpc}{Logical to use or exclude a finite population correction}
}
\value{
A tibble table combining components of a stratified estimator and providing
estimates for
\item{Ntotal}{The total population size}
\item{ntotal}{The total sample size}
\item{point}{The point estimate (stat) provided}
\item{var}{The variance of the point estimator}
\item{se}{The standard error}
\item{cv}{The coefficient of variation}
}
\description{
stratified_stat
}
\examples{
# First need to get output from make_summary
d <- data.frame("stratum"=rep(1:5, 6), "y"=rnorm(30, 30, 3))
N <- 50
ms <- make_summary(d, stratum, y, N, .stat="total")
stratified_stat(ms, .stat="total")
# Can equivalently pipe the output from make_summary into stratified_stat
# (Note how easily the estimated statistic can be computed)
make_summary(d, stratum, y, N, .stat="mean") \%>\%
stratified_stat(.stat="mean")
}
