% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sumcpp}
\alias{sumcpp}
\title{Sum of Vector Elements (C++)}
\usage{
sumcpp(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
This function uses the \code{accu()} function from the \code{Armadillo} library, a C++ library for linear algebra and
scientific computing. For \strong{double-Precision} vectors, the run time of this function is faster than the base
\code{sum()} function, making it better suited for summing the \code{weight} variable with > 3000 elements.
}
\examples{
\donttest{
# Summing a double vector
x <- runif(1000000, 0.0000001, 1.99999999)
sumcpp(x)
}
}
