% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{summarise_themes}
\alias{summarise_themes}
\title{summarise_themes}
\usage{
summarise_themes(dataframe, column, num_words = 3)
}
\arguments{
\item{dataframe}{dataframe or tibble of survey responses}

\item{column}{string variable of free text responses which has a specified LDA model}

\item{num_words}{integer denoting the number of top words to return for each theme}
}
\value{
a dataframe, where each column is a theme and contains the most common words in those themes
}
\description{
Function to return to top unique words for each theme, running
determine_themes if there is no model for that column yet
}
