% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyr-package.R
\docType{package}
\name{surveyr-package}
\alias{surveyr}
\alias{surveyr-package}
\title{surveyr: Summary Statistics and Visualizations for Surveys}
\description{
Provides tools for designing surveys and working with survey data. The package will implement methods for visualizing data, computing and tabulating summary statistics, and estimating confidence intervals for surveys using simple random samples, stratified designs, and ratio/regression estimation via the tidyverse family of packages.
}
\author{
\strong{Maintainer}: Daniel Drennan \email{danjdrennan@stat.tamu.edu}

}
\keyword{internal}
