% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_create_joint_var.R
\name{svy_create_joint_var}
\alias{svy_create_joint_var}
\title{Create a Joint Factor Variable}
\usage{
svy_create_joint_var(data, vars, sep = "/")
}
\arguments{
\item{data}{A data frame or tibble containing the survey data.}

\item{vars}{A character vector of the variable names to combine.
The order of variables matters: the levels of the first variable will vary
fastest in the resulting factor levels.}

\item{sep}{A character string to separate the levels. Default is "/".}
}
\value{
A factor vector representing the joint variable.
}
\description{
Combines multiple factor variables into a single joint factor variable.
The levels of the new factor are created by crossing the levels of the input
factors in the order specified. This is often used to create interaction
terms for raking targets (e.g., Age x Sex).
}
