% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_normalize.R
\name{svy_normalize}
\alias{svy_normalize}
\title{Normalize Weights to Sample Size}
\usage{
svy_normalize(wt_vec)
}
\arguments{
\item{wt_vec}{A numeric vector of weights.}
}
\value{
A numeric vector of normalized weights.
}
\description{
Rescales a weight vector so that it sums to the sample size (length of the vector).
This is often used to return weights to the scale of the original sample after
population projection or other adjustments.
}
