% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_projection.R
\name{svy_projection}
\alias{svy_projection}
\title{Project Weights to Population Size}
\usage{
svy_projection(wt_vec, population)
}
\arguments{
\item{wt_vec}{A numeric vector of weights.}

\item{population}{A single numeric value representing the target population size.}
}
\value{
A numeric vector of projected weights.
}
\description{
Rescales a weight vector so that it sums to a specific target population size.
}
