% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_creation.R
\name{svy_rake_optimize}
\alias{svy_rake_optimize}
\title{Optimize Raking Caps to Achieve Target Statistics}
\usage{
svy_rake_optimize(
  df,
  targets,
  max_deff = NULL,
  max_wt_ratio = NULL,
  step = 0.1,
  max_weight_start = 5,
  max_weight_min = 1.5,
  ...
)
}
\arguments{
\item{df}{A data frame or tibble containing the survey variables.}

\item{targets}{A named list of tibbles containing the population targets.}

\item{max_deff}{Numeric. The maximum allowable Design Effect (Kish). The function
will iteratively tighten the upper weight cap until this threshold is met.}

\item{max_wt_ratio}{Numeric. The maximum allowable Weight Ratio (Max/Min).
The function will iteratively tighten the caps until this threshold is met.}

\item{step}{Numeric. The step size for reducing the max_weight cap in each
iteration. Defaults to 0.1.}

\item{max_weight_start}{Numeric. The starting upper cap for weights. Defaults to 5.0.}

\item{max_weight_min}{Numeric. The floor for the upper cap (safety stop). Defaults to 1.5.}

\item{...}{Additional arguments passed to \code{svy_rake_with_trim}.}
}
\value{
A \strong{numeric vector} containing the optimized weights.
}
\description{
Automates the process of finding the best weight trimming caps to
achieve a specific Design Effect (DEFF) or Weight Ratio, while maintaining
target alignment as much as possible.
}
