% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_reproportion_targets.R
\name{svy_reproportion_targets}
\alias{svy_reproportion_targets}
\title{Reproportion Multiple Targets}
\usage{
svy_reproportion_targets(data, targets, missing_val = "Missing")
}
\arguments{
\item{data}{A data frame or tibble containing the survey data.}

\item{targets}{A named list of target data frames. The names of the list
must correspond to the variable names in the data.}

\item{missing_val}{A character string for the missing category label. Default is "Missing".}
}
\value{
A named list of reproportioned target tibbles.
}
\description{
Applies \code{svy_reproportion_target} across a list of target data frames.
}
