% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_unique.R
\name{svy_unique}
\alias{svy_unique}
\title{Get Unique Values of a Vector}
\usage{
svy_unique(x, na_rm = TRUE, sort = TRUE)
}
\arguments{
\item{x}{An atomic vector.}

\item{na_rm}{Logical. If \code{TRUE} (default), \code{NA} values are removed
from the result.}

\item{sort}{Logical. If \code{TRUE} (default), the unique values are sorted.
If \code{na_rm = FALSE} and \code{sort = TRUE}, missing values are placed at the end.}
}
\value{
A vector of the same type as \code{x} containing the unique values.
}
\description{
Returns the unique values of an atomic vector, with options to remove missing
values and sort the result.
}
