% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_topline_multiselect.R
\name{topline_caption_multiselect}
\alias{topline_caption_multiselect}
\title{Generate a list of captions for multiple selection toplines}
\usage{
topline_caption_multiselect(patterns, parent)
}
\arguments{
\item{patterns}{A character vector of "prefixes" for selecting columns in \code{df}.}

\item{parent}{A logical vector of booleans indicating whether \code{df} has a parent response column.}
}
\value{
A list of character vectors that can be passed to \code{apply_topline_multiselect()} as its \code{caption} argument.
}
\description{
This function is a helper for generating a list of captions given a \code{patterns} vector and
a \code{parent} vector. The output of this function can then be passed to \code{apply_topline_multiselect()}.
}
\examples{
\donttest{
# Vector of patterns and parent
patterns <- c("prefix_1", "predix_2", ...)
parents <- c(TRUE, FALSE, ...)

# Captions list
captions <- topline_caption_multiselect(patterns, parents)
}
}
