% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trimcpp}
\alias{trimcpp}
\title{Remove leading and trailing spaces of strings in a vector (C++)}
\usage{
trimcpp(s)
}
\arguments{
\item{s}{A character vector.}
}
\value{
A character vector.
}
\description{
This function removes all leading or trailing white space, including
space (' '), form feed (\verb{\\f}), line feed (\verb{\\n}), carriage return (\verb{\\r'},
horizontal tab (\verb{\\t}), and vertical tab (\verb{\\v}), from a character vector.
}
\examples{
\donttest{
x <- c(" leading", "trailing ", " both ")
trimcpp(x)
}
}
