% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{winfrog2dd}
\alias{winfrog2dd}
\title{Convert latitude and longitude from Winfrog format to decimal degrees.}
\usage{
winfrog2dd(x)
}
\arguments{
\item{x}{Latitude or longitude values in Winfrog format.}
}
\value{
\code{y} in decimal degrees.
}
\description{
Convert latitude and longitude from Winfrog format to decimal degrees.
}
