% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{xy2latlon}
\alias{xy2latlon}
\title{Convert x/y points to latitude/longitude}
\usage{
xy2latlon(x, y, lat, lon, units = "km")
}
\arguments{
\item{x}{X position, in meters.}

\item{y}{Y position in meters.}

\item{lat}{Latitude of known origin.}

\item{lon}{Longitude of known origin.}

\item{units}{Distance units in kilometers (default).}
}
\value{
A date/time object in POSIXct format.
}
\description{
Convert x/y points to latitude/longitude
}
