% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_xtab_args.R
\name{flatten_args}
\alias{flatten_args}
\title{Flatten list of tibble objects into a single tibble}
\usage{
flatten_args(l)
}
\arguments{
\item{l}{A list of list as returned by \code{list_xtab_args()}.}
}
\value{
A single tibble of crosstab arguments to be passed to \code{generate_tbls(type = "crosstab")}.
}
\description{
This is a helper function that accepts the output of \code{list_xtab_args()} containing two lists--- \code{return}
and \code{error}. It will either \strong{1)} throw an error indicating that attention must be paid to the \code{error}
list, or \strong{2)} return a single list of arguments, which can then be passed to \code{generate_tbls()}.
}
\examples{
\donttest{
# Generate arguments list for crosstabs
list_xtabs_args(df, var_of_interest, dependent_vars, rm) \%>\%
  flatten_args(l = .)
}
}
\seealso{
\code{\link[=list_xtab_args]{list_xtab_args()}} for vectorized crosstab arguments generation and \code{\link[=generate_tbls]{generate_tbls()}} for
multiple tables generation.
}
