% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tbls.R
\name{generate_tbls}
\alias{generate_tbls}
\title{Iteratively generate multiple tables}
\usage{
generate_tbls(l, df, weight, type = "topline")
}
\arguments{
\item{l}{A list of arguments to be passed on to either \code{generate_xtab()} or \code{generate_topline()}:
\describe{
\item{For toplines}{
\preformatted{list_topline <- tibble::tribble(
     ~x, ~caption,
     "var_1", "caption1",
     "var_2", "caption2",
     "var_3", "caption3",
     .
     .
     .
  )}}
\item{For crosstabs}{
\preformatted{list_xtab <- tibble::tribble(
     ~x, ~y, ~caption,
     "var_1", "var_2" "caption1",
     "var_2", "var_9", caption2",
     "var_3", "var_23", "caption3",
     .
     .
     .
  )}}
}}

\item{df}{A data frame or tibble.}

\item{weight}{A length one character vector used as the caption for the topline.}

\item{type}{Must either be 'topline' or 'crosstab'. Defaults to 'topline'.}
}
\value{
A list containing elements that are length one character vectors, each of which is the source code for a single
table.
}
\description{
This function provides a wrapper around \code{generate_xtab()} and \code{generate_topline()}, allowing users
to efficiently generate many tables by passing a list of arguments. The shapes of the argument lists
may vary depending on the table \code{type}.
}
\examples{
# Generate crosstabs
list_of_xtabs <- generate_tbls(
  l = list_xtab,
  df,
  "weight",
  "crosstab"
)

# Generate toplines
list_of_topline <- generate_tbls(
  l = list_topline,
  df,
  "weight",
  "topline"
)
}
\seealso{
\code{\link[=generate_xtab]{generate_xtab()}} and \code{\link[=generate_topline]{generate_topline()}} for single crosstab or topline generation.
}
