% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{graph_responses}
\alias{graph_responses}
\title{creates a stacked, single-bar chart showing the fraction of
responses for each option (ie, the column totals 100% (% of responses))}
\usage{
graph_responses(mytable, mytitle)
}
\arguments{
\item{mytable}{the output of process_radio_button_q()}

\item{mytitle}{Title text for the graph}
}
\value{
a ggplot object
}
\description{
creates a stacked, single-bar chart showing the fraction of
responses for each option (ie, the column totals 100% (% of responses))
}
