% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{graph_responses_by_grp}
\alias{graph_responses_by_grp}
\title{graph_responses_by_grp: creates a stacked, multicolumn bar chart where
each bar represents one group in 'mygroup'; each column totals 100% of
responses in that group}
\usage{
graph_responses_by_grp(mytable, mygroup, mytitle)
}
\arguments{
\item{mytable}{the output of process_radio_button_q_by_grp()}

\item{mygroup}{the column whose values define a group}

\item{mytitle}{graph title}
}
\value{
a ggplot object.
}
\description{
graph_responses_by_grp: creates a stacked, multicolumn bar chart where
each bar represents one group in 'mygroup'; each column totals 100% of
responses in that group
}
