% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lln_demo.R
\name{lln_demo}
\alias{lln_demo}
\title{Law of Large Numbers Demo}
\usage{
lln_demo(n = 5000)
}
\arguments{
\item{n}{A positive integer greater than 100.}
}
\value{
A list with
\item{data}{The constructed dataset used for plotting}
\item{plot}{A ggplot plot object showing convergence of the mean and variance}
}
\description{
Provides a demonstration of the law of large numbers using draws from a
gamma distribution with shape 20 and scale 1/20.
The mean and variance, then, are \eqn{k\theta = 120} and
\eqn{k\theta^2 = 2400} for each i.
Draws from the distribution used here can be made by \code{rgamma(n, 6, 1/20)};
see \code{?rgamma} for more details about the function.

The law of large numbers essentially states that, for independent samples
from some parent distribution, the sample mean (variance) converges in
probability towards the expected value of a population statistic. More precise
statements can be found in the references or from Wikipedia.

The Law of Large Numbers is useful because it gives conditions under which
an estimator converges to its population value.
}
\examples{
lln_demo()

}
\references{
Billingsley, Patrick. Probability and measure. John Wiley & Sons, 2008.
}
