% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.R
\name{prettify}
\alias{prettify}
\title{Function which takes in a formattable and abstracts the use of the formattable package by
creating a simple, standardised output}
\usage{
prettify(object, alias = c(""), align = c("l"), count_bar = FALSE,
  colour_groups = FALSE, flatten = FALSE)
}
\arguments{
\item{object}{formattable object created by common_words function}

\item{alias}{character vector indicating what the columns should be renamed, must be a named vector or
the same length as the number of columns}

\item{count_bar}{logical indicating whether a bar should be added to the count column to visualise the count}

\item{colour_groups}{logical indicating whether the groups of the first column should be distinguished by colours}
}
\value{
Newly formatted formattable object
}
\description{
Function which takes in a formattable and abstracts the use of the formattable package by
creating a simple, standardised output
}
