% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print_tbls}
\alias{print_tbls}
\title{Print tables iteratively}
\usage{
print_tbls(l, output)
}
\arguments{
\item{l}{A list of \code{flextable} or \code{kableExtra} objects, which are s3 lists with certain structures.}

\item{output}{A string specifying the output format.}
}
\value{
Invisible.
}
\description{
This function takes a list of \code{flextable} or \code{kableExtra} objects and prints the raw code adapted to the 'output'
format. \strong{Note} that chunk option results must be set to \strong{asis}, i.e., \code{results='asis'}.
}
\examples{
\donttest{
# Toplines to word
print_tbls(list_toplines, output = "word")

# Crosstabs to pdf
print_tbls(list_xtabs, output = "pdf")
}
}
