% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{process_free_text_q}
\alias{process_free_text_q}
\title{Count regex terms in free-text answers. Multiple matches in
a single answer are counted once.}
\usage{
process_free_text_q(mytable, question, terms)
}
\arguments{
\item{mytable}{the dataset}

\item{question}{question under analysis (use the janitor::clean_names() text)}

\item{terms}{a vector of regexes to search for}
}
\value{
a tibble tallying the occurrence of each term in 'terms' in the question
}
\description{
Count regex terms in free-text answers. Multiple matches in
a single answer are counted once.
}
