% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{process_multiq_radio_button_q}
\alias{process_multiq_radio_button_q}
\title{process_multiq_radio_button_q: Processes a multi-question radio button}
\usage{
process_multiq_radio_button_q(mytable, question, mylevels)
}
\arguments{
\item{mytable}{the dataset}

\item{question}{question under analysis (use the janitor::clean_names() text)}

\item{mylevels}{vector of possible answers to the question}
}
\value{
a list of tibbles, each named according to the question, each with cols 'option', 'count', 'pct'
}
\description{
process_multiq_radio_button_q: Processes a multi-question radio button
}
