% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{process_multiq_radio_button_q_by_grp}
\alias{process_multiq_radio_button_q_by_grp}
\title{process_multiq_radio_button_q_by_grp: Processes a multi-question radio button question,
but breaks down numbers by group}
\usage{
process_multiq_radio_button_q_by_grp(mytable, mygroup, question, mylevels)
}
\arguments{
\item{mytable}{the dataset}

\item{mygroup}{the column whose values define a group}

\item{question}{question under analysis (use the janitor::clean_names() text)}

\item{mylevels}{vector of possible answers to the question}
}
\value{
a list of tibbles, each named for one question in the overall question
and containing cols 'option', <mygroup>, 'count', 'pct'
}
\description{
process_multiq_radio_button_q_by_grp: Processes a multi-question radio button question,
but breaks down numbers by group
}
