% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{process_radio_button_q}
\alias{process_radio_button_q}
\title{process_radio_button_q: Process a standard radio button question}
\usage{
process_radio_button_q(mytable, question, mylevels)
}
\arguments{
\item{mytable}{the dataset}

\item{question}{question under analysis (use the janitor::clean_names() text)}

\item{mylevels}{vector of possible answers to the question}
}
\value{
tibble with cols 'option', 'count', 'pct', where
count = count of responses, pct = count as a fraction of all responses (count / sum(count))
}
\description{
process_radio_button_q: Process a standard radio button question
}
