% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_distribution.R
\name{sampling_distribution}
\alias{sampling_distribution}
\title{sampling_distribution}
\usage{
sampling_distribution(y, n)
}
\arguments{
\item{y}{A numeric vector of measurements on observations units.
The population, N, will be computed as the length of y.}

\item{n}{The size of samples to draw from y.}
}
\value{
The sampling distribution of the sample mean
}
\description{
Computes the sampling distribution of an estimator given a population,
sample size, and statistic.
The population must be a vector of numeric variables.
The statistic must be the sample mean, sample total, or sample proportion.
The function must store the entire sample as a matrix at some point, so it is
bound on what samples can taken by choose(N, n).
}
\examples{
y <- c(1, 1, 2, 2, 3, 3)
n <- 2
sampling_distribution(y, n)
# 1.0 1.5 1.5 2.0 2.0 1.5 1.5 2.0 2.0 2.0 2.5 2.5 2.5 2.5 3.0
}
