% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_reproportion_targets.R
\name{svy_reproportion_targets}
\alias{svy_reproportion_targets}
\title{Reproportion Targets to Account for Missing Data}
\usage{
svy_reproportion_targets(df, targets, missing_cat = "Missing")
}
\arguments{
\item{df}{The dataset (data.frame or tibble).}

\item{targets}{A named list of data frames representing target distributions.
The list names must correspond to variables in \code{df}.}

\item{missing_cat}{Character string for the missing category label. Default is 'Missing'.}
}
\value{
A modified list of targets.
}
\description{
Adjusts target distributions to include a category for missing values (NA) found in the
sample data. This ensures that records with missing data on weighting variables are
accounted for (assigned a specific target proportion) rather than being dropped or
causing errors during raking.
}
